requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cattrs==1.6.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
dateparser==1.0.0
facebook-scraper==0.2.35
facebook-sdk==3.1.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
parse==1.19.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-facebook-api==0.9.2
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
responses==0.13.3
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.14
tqdm==4.60.0
tzlocal==2.1
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
